+++ astropy/units/equivalencies.py
@@ -13,6 +13,7 @@ from . import cgs
 from . import astrophys
 from .function import units as function_units
 from . import dimensionless_unscaled
+from .core import UnitsError
 
 
 __all__ = ['parallax', 'spectral', 'spectral_density', 'doppler_radio',
@@ -208,6 +209,8 @@ def doppler_radio(rest):
     <Quantity -31.209092088877583 km / s>
     """
 
+    assert_is_spectral_unit(rest)
+
     ckms = _si.c.to('km/s').value
 
     def to_vel_freq(x):
@@ -272,6 +275,8 @@ def doppler_optical(rest):
     >>> optical_velocity  # doctest: +FLOAT_CMP
     <Quantity -31.20584348799674 km / s>
     """
+
+    assert_is_spectral_unit(rest)
 
     ckms = _si.c.to('km/s').value
 
@@ -345,6 +350,8 @@ def doppler_relativistic(rest):
     >>> relativistic_wavelength  # doctest: +FLOAT_CMP
     <Quantity 2.6116243681798923 mm>
     """
+
+    assert_is_spectral_unit(rest)
 
     ckms = _si.c.to('km/s').value
 
@@ -479,3 +486,10 @@ def temperature_energy():
     return [
         (si.K, si.eV, lambda x: x / (_si.e.value / _si.k_B),
          lambda x: x * (_si.e.value / _si.k_B))]
+
+def assert_is_spectral_unit(value):
+    try:
+        value.to(si.Hz, spectral())
+    except (AttributeError, UnitsError) as ex:
+        raise UnitsError("The 'rest' value must be a spectral equivalent "
+                         "(frequency, wavelength, or energy).")