+++ astropy/units/tests/test_equivalencies.py
@@ -71,17 +71,17 @@ def test_logarithmic(log_unit):
                       1.e-10*log_unit)
 
 
-functions = [u.doppler_optical, u.doppler_radio, u.doppler_relativistic]
+doppler_functions = [u.doppler_optical, u.doppler_radio, u.doppler_relativistic]
 
 
-@pytest.mark.parametrize(('function'), functions)
+@pytest.mark.parametrize(('function'), doppler_functions)
 def test_doppler_frequency_0(function):
     rest = 105.01 * u.GHz
     velo0 = rest.to(u.km/u.s, equivalencies=function(rest))
     assert velo0.value == 0
 
 
-@pytest.mark.parametrize(('function'), functions)
+@pytest.mark.parametrize(('function'), doppler_functions)
 def test_doppler_wavelength_0(function):
     rest = 105.01 * u.GHz
     q1 = 0.00285489437196 * u.m
@@ -89,7 +89,7 @@ def test_doppler_wavelength_0(function):
     np.testing.assert_almost_equal(velo0.value, 0, decimal=6)
 
 
-@pytest.mark.parametrize(('function'), functions)
+@pytest.mark.parametrize(('function'), doppler_functions)
 def test_doppler_energy_0(function):
     rest = 105.01 * u.GHz
     q1 = 0.000434286445543 * u.eV
@@ -97,7 +97,7 @@ def test_doppler_energy_0(function):
     np.testing.assert_almost_equal(velo0.value, 0, decimal=6)
 
 
-@pytest.mark.parametrize(('function'), functions)
+@pytest.mark.parametrize(('function'), doppler_functions)
 def test_doppler_frequency_circle(function):
     rest = 105.01 * u.GHz
     shifted = 105.03 * u.GHz
@@ -106,7 +106,7 @@ def test_doppler_frequency_circle(function):
     np.testing.assert_almost_equal(freq.value, shifted.value, decimal=7)
 
 
-@pytest.mark.parametrize(('function'), functions)
+@pytest.mark.parametrize(('function'), doppler_functions)
 def test_doppler_wavelength_circle(function):
     rest = 105.01 * u.nm
     shifted = 105.03 * u.nm
@@ -115,7 +115,7 @@ def test_doppler_wavelength_circle(function):
     np.testing.assert_almost_equal(wav.value, shifted.value, decimal=7)
 
 
-@pytest.mark.parametrize(('function'), functions)
+@pytest.mark.parametrize(('function'), doppler_functions)
 def test_doppler_energy_circle(function):
     rest = 1.0501 * u.eV
     shifted = 1.0503 * u.eV
@@ -126,13 +126,20 @@ def test_doppler_energy_circle(function):
 
 values_ghz = (999.899940784289,999.8999307714406,999.8999357778647)
 @pytest.mark.parametrize(('function', 'value'),
-                         list(zip(functions, values_ghz)))
+                         list(zip(doppler_functions, values_ghz)))
 def test_30kms(function, value):
     rest = 1000 * u.GHz
     velo = 30 * u.km/u.s
     shifted = velo.to(u.GHz, equivalencies=function(rest))
     np.testing.assert_almost_equal(shifted.value, value, decimal=7)
 
+
+bad_values = (5, 5*u.Jy, None)
+@pytest.mark.parametrize(('function', 'value'),
+                         list(zip(doppler_functions, bad_values)))
+def test_bad_restfreqs(function, value):
+    with pytest.raises(u.UnitsError):
+        function(value)
 
 def test_massenergy():
     # The relative tolerance of these tests is set by the uncertainties